+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 This project follows [Semantic Versioning](http://semver.org/).
 
+## [8.0.7] 2022-12-14
+### Added
+- Support import of forms that are in LForms format.
+
+## [8.0.6] 2022-12-13
+### Fixed
+- Fix display of decimal answer field in conditional display when source item is decimal type.
+
+## [8.0.5]  2022-12-05
+### Fixed
+- Fix loading initial component when switched type from choice to decimal.
+
 ## [8.0.4] 2022-11-28
 ### Fixed
 - Fix placing cursor position in integer input when invalid character is typed.